products/b-manager/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
line-bot-sdk>=3.14.0
anthropic>=0.52.0
google-api-python-client>=2.170.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.1.0
apscheduler>=3.11.0
httpx>=0.28.0
